fix merged tests
