+++ spring-cloud-kubernetes-config/src/test/java/org/springframework/cloud/kubernetes/config/KubernetesConfigConfigurationTest.java
@@ -76,43 +76,6 @@ public class KubernetesConfigConfigurationTest extends KubernetesConfigTestBase 
 	}
 
 	@Test
-	public void kubernetesReloadEnabled() throws Exception {
-		setup("spring.cloud.kubernetes.enabled=true",
-				"spring.cloud.kubernetes.reload.enabled=true");
-		assertThat(this.getContext().containsBean("configMapPropertySourceLocator"))
-				.isTrue();
-		assertThat(this.getContext().containsBean("secretsPropertySourceLocator"))
-				.isTrue();
-		assertThat(this.getContext().containsBean("propertyChangeWatcher")).isTrue();
-	}
-
-	@Test
-	public void kubernetesReloadEnabledButSecretDisabled() throws Exception {
-		setup("spring.cloud.kubernetes.enabled=true",
-				"spring.cloud.kubernetes.config.enabled=true",
-				"spring.cloud.kubernetes.secrets.enabled=false",
-				"spring.cloud.kubernetes.reload.enabled=true");
-		assertThat(this.getContext().containsBean("configMapPropertySourceLocator"))
-				.isTrue();
-		assertThat(this.getContext().containsBean("secretsPropertySourceLocator"))
-				.isFalse();
-		assertThat(this.getContext().containsBean("propertyChangeWatcher")).isTrue();
-	}
-
-	@Test
-	public void kubernetesReloadEnabledButSecretAndConfigDisabled() throws Exception {
-		setup("spring.cloud.kubernetes.enabled=true",
-				"spring.cloud.kubernetes.config.enabled=false",
-				"spring.cloud.kubernetes.secrets.enabled=false",
-				"spring.cloud.kubernetes.reload.enabled=true");
-		assertThat(this.getContext().containsBean("configMapPropertySourceLocator"))
-				.isFalse();
-		assertThat(this.getContext().containsBean("secretsPropertySourceLocator"))
-				.isFalse();
-		assertThat(this.getContext().containsBean("propertyChangeWatcher")).isFalse();
-	}
-
-	@Test
 	public void kubernetesConfigwhenKubenretesEnabledAndKubernetsConfigEnabled()
 			throws Exception {
 		setup("spring.cloud.kubernetes.config.enabled=true",
